Add config setting to all PHP files
All php files now require a config file to e included. This makes life
easier for us as we can all have own passwords/usernames/database names
etc.

The config file is not tracked by git so you'll have to copy the shared
one and edit it. Put it in the same folder as the php scripts.

Mae sure all new php files use this method as well.